+++ components/SocialIcons.tsx
@@ -1,11 +1,12 @@
 import clsx from 'clsx';
-import { Colors, contact, ContactType } from 'config';
+import { Colors } from 'config/colors';
+import { contact, ContactType } from 'config/contact';
 import React, { memo } from 'react';
 import { IconBaseProps } from 'react-icons';
 import { FaLinkedinIn as LinkedinIcon } from 'react-icons/fa';
 import {
   IoLogoGithub as GithubIcon,
-  IoLogoTwitter as TwitterIcon
+  IoLogoTwitter as TwitterIcon,
 } from 'react-icons/io';
 import { SiGmail as MailIcon } from 'react-icons/si';
 import { Maybe, Tuple } from 'types';
