+++ config/colors.ts
@@ -1,0 +1,48 @@
+export const Colors: Record<string, string> = {
+  // Languages
+  go: '#00ADD8',
+  python: '#4B8BBE',
+  typescript: '#234A84',
+  javascript: '#F7DF1E',
+
+  // Frontend
+  react: '#61DAF6',
+  nextjs: '#000000',
+
+  // Backend
+  graphql: '#E535AB',
+  node: '#68A063',
+  django: '#092E20',
+
+  // Tools, Libs
+  babel: '#F5DB53',
+  redux: '#764ABC',
+
+  // Mobile
+  reactnative: '#2D2D2D',
+  android: '#56A036',
+  ios: '#0C76E2',
+
+  // Databases
+  postgres: '#336791',
+  mongo: '#4DB33D',
+  redis: '#D82C20',
+  webpack: '#8DD6F9',
+
+  // Cloud
+  aws: '#FF9900',
+  gcp: '#4285F4',
+  docker: '#0DB7Ed',
+  kubernetes: '#326CE5',
+  terraform: '#7B42BC',
+
+  // Social
+  linkedin: '#0077B5',
+  twitter: '#1DA1F2',
+  email: '#D44638',
+
+  // Misc.
+  testing: '#049C64',
+  devops: '#059F00',
+  resume: '#EEEBFF',
+};
